% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjebutils.R
\name{anonymize}
\alias{anonymize}
\title{Extract MAIA raw data}
\usage{
anonymize(df)
}
\value{
Data frame
}
\description{
The function will anonymize your data
}
\author{
Stackoverflow::Etienne Low-Décarie, slightly modified
}
