% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjebutils.R
\name{count_pateyes}
\alias{count_pateyes}
\title{count_pateyes}
\usage{
count_pateyes(y, text = FALSE)
}
\arguments{
\item{y}{dataframe where patient / eye information is included.
The patient ID column name needs to contain the strings "pat" OR "id" ignore.case = TRUE.
The eye column name needs to contain the string "eye" ignore.case = TRUE}

\item{return_text}{default = FALSE (TRUE will return text which can be pasted for example into a markdown document). As default a named vector will be returned}
}
\description{
Looks for columns that identify patients and eyes and counts number of patients and eyes.
}
