% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjebutils.R
\name{csv}
\alias{csv}
\title{csv}
\usage{
csv(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{data frame}

\item{name}{name of csv file (does not need .csv extension), (quoted) string. If not specified, the name of the data frame will be used}
}
\description{
wrapper around write.csv with default 'row.names = FALSE'
}
