% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_week.R
\name{days}
\alias{days}
\title{names of weekdays}
\usage{
days(index = 1, abbreviate = FALSE)
}
\arguments{
\item{index}{with which day in the week 0:6 shall the vector start,
d=0 is Sunday, d=1 is Monday, ..., d=6 is Saturday}

\item{abbreviate}{logical vector (possibly recycled).
Should the names be abbreviated?}
}
\description{
a convenient function to create a vector of weekdays
in the locale
}
