% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{every_nth}
\alias{every_nth}
\title{every_nth}
\usage{
every_nth(x, nth, empty = TRUE, inverse = FALSE)
}
\arguments{
\item{x}{vector}

\item{nth}{how many gaps}

\item{empty}{how to fill}

\item{inverse}{if TRUE removes every nth instead}
}
\value{
char vector
}
\description{
creates vector with empty lables for custom major/minor ticks
}
\seealso{
https://stackoverflow.com/a/34533473/7941188
}
\author{
Stackoverflow user adamdsmith
}
