% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjebutils.R
\name{find_object}
\alias{find_object}
\title{find_object}
\usage{
find_object(df, col_name, list)
}
\description{
looks up content of column in list and creates new column assigning corresponding characteristic to list
}
\author{
tjebo
}
