% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_type_b.R
\name{geom_pointpath}
\alias{geom_pointpath}
\title{geom_pointpath}
\usage{
geom_pointpath(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
makes point and line plot just as plot type = b
}
