% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwet_ac1_raw.R
\name{identity.weights}
\alias{identity.weights}
\title{making identity weights}
\usage{
identity.weights(categ)
}
\description{
needed for all IRR computationts from the raredd/ragree package
}
\references{
Gwet, K. L. (2008). "Computing inter-rater reliability and its
variance in the presence of high agreement." British Journal of Mathematical
and Statistical Psychology, 61, 29 - 48.
}
\author{
Kilem L. Gwet
}
