% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivoting.R
\name{longer}
\alias{longer}
\title{pivoting}
\usage{
longer(x, cols, names_to, values_to, ...)
}
\arguments{
\item{cols}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns to pivot into
longer format.}

\item{names_to}{A string specifying the name of the column to create
from the data stored in the column names of \code{data}.

Can be a character vector, creating multiple columns, if \code{names_sep}
or \code{names_pattern} is provided. In this case, there are two special
values you can take advantage of:
\itemize{
\item \code{NA} will discard that component of the name.
\item \code{.value} indicates that component of the name defines the name of the
column containing the cell values, overriding \code{values_to}.
}}

\item{values_to}{A string specifying the name of the column to create
from the data stored in cell values. If \code{names_to} is a character
containing the special \code{.value} sentinel, this value will be ignored,
and the name of the value column will be derived from part of the
existing column names.}

\item{...}{Additional arguments passed on to methods.}
}
\value{
data frame
}
\description{
My own twist on tidyr::pivot_longer and tidyr::pivot_wider,
just for convenience
}
