% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_type_b.R
\name{makeContent.pointpath}
\alias{makeContent.pointpath}
\title{makeContent.pointpath}
\usage{
\method{makeContent}{pointpath}(x)
}
\description{
underlying function for geom_type_b
}
\author{
SO::teunbrand
}
