% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{match_part}
\alias{match_part}
\alias{match_replace}
\title{Partial string matching and replacing}
\usage{
match_part(x, table)

match_replace(x, replace, no_match = NA_character_)
}
\arguments{
\item{x}{Vector. Values to be matched}

\item{table}{Vector. characters to be matched against}

\item{replace}{Vector of new labels if a partial match is found.
If named vector, names are used. If more than one match is found,
the last match will replace the original value.}

\item{no_match}{if there is no partial match}
}
\value{
A list of indices of partial matches

Character vector with replaced values
}
\description{
Partial string matching and replacing

Partial string matching and replacing
}
\examples{
oceans <- c("pacific", "atlantic", "indian", "mediterranean")
oceans_long <- c("Eastern Central Atlantic", "Eastern Central Pacific", "Eastern Indian Ocean",
  "Mediterranean and Black Sea", "Northeast Central Atlantic",
  "Northeast Pacific", "Northwest Central Atlantic", "Northwest Pacific",
  "Southeast Central Atlantic", "Southeast Pacific", "Southwest Atlantic",
  "Southwest Pacific", "Western Central Atlantic", "Western Central Pacific",
  "Western Indian Ocean", "World", "Atlantic Pacific")

match_replace(oceans_long, oceans, "Other")
}
