% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mysplit}
\alias{mysplit}
\title{csv}
\usage{
mysplit(x, col)
}
\arguments{
\item{x}{data frame}

\item{col}{column to split by}
}
\value{
list
}
\description{
Because I am using split so often, in order to split by
a single column.
}
\examples{
mysplit(mtcars, cyl)
}
\seealso{
Other convenience functions: 
\code{\link{csv}()},
\code{\link{str_title}()}
}
\concept{convenience functions}
