% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_week.R
\name{next_weekday}
\alias{next_weekday}
\title{Next weekday to a given date}
\usage{
next_weekday(x, index, method = "ceiling")
}
\arguments{
\item{x}{date vector}

\item{index}{which day in the week 0:6,
d=0 is Sunday, d=1 is Monday, ..., d=6 is Saturday}

\item{method}{"ceiling": last day of the week given the weekend.
"floor": first day of week}
}
\description{
Gives the date of a given weekday to a given date
}
\details{
The idea is that relative to the UNIX Epoch that d-4 falls on day of
the week d where d=0 is Sunday, d=1 is Monday, ..., d=6 is Saturday
so any multiple of 7 days from that also falls on day of the week d.
}
