% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{prob_contour}
\alias{prob_contour}
\title{Probability contours}
\usage{
prob_contour(data, x = NULL, y = NULL, n = 50, prob = 0.95, ...)
}
\arguments{
\item{x}{data frame with two columns (x and y coordinates)}

\item{name}{Filename. Default: Name of dataframe to save as csv. Or character string (.csv extension added automatically)}
}
\description{
calculates 2d probability contours for use in ggplot2
}
\examples{

set.seed(1)
n=100
foo <- data.frame(x=rnorm(n, 0, 1), y=rnorm(n, 0, 1))

df_contours <- dplyr::bind_rows(
  purrr::map(seq(0.2, 0.8, 0.2), function(p) prob_contour(foo, prob = p))
)

ggplot() +
  stat_density_2d(data = foo, aes(x, y), bins = 5, color = "black") +
  geom_point(data = foo, aes(x = x, y = y)) +
  geom_polygon(data = df_contours, aes(x = x, y = y, color = prob), fill = NA) +
  scale_color_brewer(name = "Probs", palette = "Set1")
}
