% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_funs.R
\name{se_diff_means}
\alias{se_diff_means}
\title{se_diff_means}
\usage{
se_diff_means(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{sample one - as numeric vector}

\item{y}{sample two}

\item{na.rm}{passed to sd}
}
\value{
numeric vector of length 1
}
\description{
The function will compute the pooled standard error of the
differences of the means between two independent samples,
with equal variances (!)
}
\seealso{
https://stats.stackexchange.com/questions/302445/standard-error-of-difference
https://www.bmj.com/about-bmj/resources-readers/publications/statistics-square-one/7-t-tests
}
