% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_funs.R
\name{se_mean}
\alias{se_mean}
\title{standard error (of the mean)}
\usage{
se_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{sample one}

\item{na.rm}{NA removed}
}
\value{
numeric vector of length 1
}
\description{
Gets the standard error of the mean of a sample
}
\seealso{
https://stackoverflow.com/a/7220087/7941188
}
