% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs_dependencies.R
\name{show_pkg_versions}
\alias{show_pkg_versions}
\title{Show current package dependencies}
\usage{
show_pkg_versions(
  pkg = ".",
  fields = c("Depends", "Imports", "LinkingTo", "Suggests")
)
}
\arguments{
\item{pkg}{package description, can be path or package name}

\item{fields}{fields to get & report dependencies for}
}
\description{
Will \code{cat} out a cut/paste-able set of fields for a
\code{DESCRIPTION} file with minimum required versions for
each package based upon currently available package vesions
in CRAN.
}
\note{
R and the R version is NOT added to \code{Depends}
}
\examples{
\dontrun{
show_pkg_versions("qmethod")
show_pkg_versions("MASS")
# assumes you're in a pkg devel dir
show_pkg_versions()
}
}
\seealso{
https://stackoverflow.com/a/38743324/7941188
}
