% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjebutils.R
\name{show_stats}
\alias{show_stats}
\title{show_stats}
\usage{
show_stats(x)
}
\arguments{
\item{x}{numeric/integer vector}
}
\value{
named vector
}
\description{
pulls the most commonly used summary statistics
}
