% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_week.R
\name{show_week}
\alias{show_week}
\title{Show ceiling / floor days of a week}
\usage{
show_week(x, week_end = "sunday", ceiling = TRUE)
}
\arguments{
\item{x}{Vector of dates Dataframe in which the vectors are found}

\item{week_end}{what's the end of week? Takes english names of weekdays as quoted string, capital letters allowed}

\item{ceiling}{default TRUE: last day of the week for each weekend. if FALSE, first day of week}
}
\value{
vector of dates. can be used in aggregations, e.g. with zoo.aggregate
}
\description{
Takes vector of dates and outputs floor or ceiling date of the respective week.
The end of week can be defined. The output can be used e.g. for zoo aggregation.
This is inspired tremendously by Stackoverflow users G.Grothendieck and Joshua Ulrich
}
\seealso{
https://stackoverflow.com/a/4351626/7941188 and
https://stackoverflow.com/a/16193667/7941188
}
