% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{str_title}
\alias{str_title}
\title{Title case}
\usage{
str_title(x, strict = FALSE, except = NULL)
}
\arguments{
\item{x}{string to convert to title case}

\item{strict}{FALSE: Abbreviations are kept all CAPS -
in fact, all letters after the first are kept in their original case}

\item{except}{Exceptions to be kept in lower case. By default:
except <- c("and", "or", "nor", "but", "a", "an", "the", "as", "at",
"by", "for", "in", "of", "on", "per", "to")}
}
\value{
string
}
\description{
Converting to title case in a clever way
}
\examples{
str_title(c("using AIC for model selection"))
}
\seealso{
\link{toupper}

Other convenience functions: 
\code{\link{csv}()},
\code{\link{mysplit}()}
}
\concept{convenience functions}
