\name{tjebtools-package}
\alias{tjebtools-package}
\alias{tjebtools}
\docType{package}
\title{
\packageTitle{tjebtools}
}
\description{
\packageDescription{tjebtools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tjebtools}
\packageIndices{tjebtools}
~~ Some more or less useful functions which help me in my day to day work ~~
}
\author{
\packageAuthor{tjebtools}

Maintainer: \packageMaintainer{tjebtools}
}
\references{
~~ TBC ~~
}
~~ TBC ~~
\keyword{ package }

\examples{
~~ TBC ~~
}
