% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{uk_bankholidays}
\alias{uk_bankholidays}
\title{UK_bankholidays}
\usage{
uk_bankholidays(year = timeDate::getRmetricsOptions("currentYear"))
}
\arguments{
\item{year}{an integer value or vector of years, formatted as YYYY.}
}
\value{
vector of POSIX.
}
\description{
modified function timeDate::holidayLONDON -
for description see ?timeDate::holidayLONDON

includes names of bank holidays in output vector
}
