% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjebutils.R
\name{weekly}
\alias{weekly}
\title{weekly}
\usage{
weekly(x, week_end = "sunday")
}
\arguments{
\item{x}{Vector of dates Dataframe in which the vectors are found}

\item{week_end}{desired end of week}
}
\value{
vector of dates. can be used in aggregations, e.g. with zoo.aggregate
}
\description{
makes vector of dates of beginning of full (!) weeks
}
\author{
tjebo - eclecticistic choice of functions, mainly with the help of SO::G.Grothendieck and SO::Joshua Ulrich
https://stackoverflow.com/a/4351626/7941188 and https://stackoverflow.com/a/16193667/7941188
}
