% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivoting.R
\name{wider}
\alias{wider}
\title{pivoting}
\usage{
wider(x, names_from, values_from, ...)
}
\arguments{
\item{names_from}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> A pair of
arguments describing which column (or columns) to get the name of the
output column (\code{names_from}), and which column (or columns) to get the
cell values from (\code{values_from}).

If \code{values_from} contains multiple values, the value will be added to the
front of the output column.}

\item{values_from}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> A pair of
arguments describing which column (or columns) to get the name of the
output column (\code{names_from}), and which column (or columns) to get the
cell values from (\code{values_from}).

If \code{values_from} contains multiple values, the value will be added to the
front of the output column.}

\item{...}{Additional arguments passed on to methods.}
}
\value{
data frame
}
\description{
My own twist on tidyr::pivot_longer and tidyr::pivot_wider,
just for convenience
}
