% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjebutils.R
\name{count_pateyes}
\alias{count_pateyes}
\title{count_pateyes}
\usage{
count_pateyes(y, return_vec = FALSE)
}
\arguments{
\item{y}{dataframe where patient / eye information is included. column which contains the patient ID must include the strings "pat" and "id". the column which contains eye must contain the string "eye". There should be only one column which fulfills each condition.}

\item{return_vec}{default = FALSE (will return text which can be pasted for example into a markdown document). If TRUE, a named vector will be returned}
}
\description{
Looks for columns that identify patients and eyes and counts number of patients and eyes.
}
