% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjebutils.R
\name{every_nth}
\alias{every_nth}
\title{every_nth}
\usage{
every_nth(x, nth, empty = TRUE, inverse = FALSE)
}
\value{
char vector
}
\description{
creates vector with empty lables for custom major/minor ticks
}
\author{
SO::adamdsmith
}
