% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjebutils.R
\name{show_stats}
\alias{show_stats}
\title{show_stats}
\usage{
show_stats(x, dec = 1)
}
\arguments{
\item{x}{either vector or list of vectors. Data frames supported. Character list elements (or columns) are removed!!}

\item{dec}{how many decimals are displayed}
}
\value{
named vector (for vector) or data frame (for list)
}
\description{
pulls the most commonly used summary statistics
}
