% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjebutils.R
\name{uk_bankholidays}
\alias{uk_bankholidays}
\title{UK_bankholidays}
\usage{
uk_bankholidays(year = getRmetricsOptions("currentYear"))
}
\value{
vector of POSIX.
}
\description{
includes NAMES of bank holidays in output vector
}
\author{
modified function timeDate::holidayLONDON - for description see ?timeDate::holidayLONDON
}
