% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_week.R
\name{weeks}
\alias{weeks}
\alias{week_ceiling}
\alias{week_floor}
\alias{days}
\title{Show ceiling / floor days of a week}
\usage{
week_ceiling(x, weekend = "sunday", ceiling = TRUE)

week_floor(x, weekend = "sunday", ceiling = FALSE)

days(abbreviate = FALSE)
}
\arguments{
\item{x}{Vector of dates}

\item{weekend}{what's the end of week? english (!) weekday names
as character, cases are ignored}

\item{ceiling}{default TRUE: last day of the week for each weekend. if FALSE, first day of week}

\item{abbreviate}{logical vector (possibly recycled).
Should the names be abbreviated?}
}
\value{
vector of dates. can be used in aggregations, e.g. with zoo.aggregate
}
\description{
Takes vector of dates and outputs floor or ceiling date of the respective week.
The end of week can be defined. The output can be used e.g. for zoo aggregation.
This is inspired tremendously by Stackoverflow users G.Grothendieck and Joshua Ulrich

a convenient function to create a vector of weekdays
in the locale
}
\examples{
week_ceiling(as.Date("2021-03-24"))
week_ceiling(as.Date("2021-03-24"), weekend = "friday")
week_floor(as.Date("2021-03-24"))
}
\seealso{
https://stackoverflow.com/a/4351626/7941188 and
https://stackoverflow.com/a/16193667/7941188
}
